% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate_gif}
\alias{animate_gif}
\title{Create a gif animation using rasterVis::levelplot() and ImageMagick's convert}
\usage{
animate_gif(
  R,
  ...,
  mains = names(R),
  other_layers = NULL,
  width = 480,
  height = 480,
  output_filename = file.path(".", "animation.gif"),
  tmppath = tempfile(pattern = "animation"),
  convert_extra = "-loop 0 -delay 50"
)
}
\arguments{
\item{R}{a multi-layer raster object}

\item{...}{further arguments for rasterVis::levelplot like
 \itemize{
   \item{margin}
   \item{at}
   \item{par.settings}
}}

\item{mains}{character vector of names to apply to each plot, by default the layer names}

\item{other_layers}{a list of one or more latticeExtra::layer or list of layers}

\item{width}{numeric image width in pixels}

\item{height}{image height in pixels}

\item{output_filename}{the name of the output file, by default 'animation.gif'}

\item{tmppath}{string, temporary working directory to use}

\item{convert_extra}{character of extra conver arguments.  Defaults to
"-loop 0 -delay = 50" for infinte looping and 0.5s per frame}
}
\value{
the numeric value returned by convert
}
\description{
This is Unix/Linux centric.
}
\seealso{
\url{http://www.imagemagick.org/Usage/anim_basics/}

\url{http://imagemagick.org/script/command-line-processing.php}
}
