% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{as_POSIXct}
\alias{as_POSIXct}
\title{Convert a Date, character, or POSIXlt to POSIXct}
\usage{
as_POSIXct(x, time = "00:00:00", tz = "UTC",
  format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{x}{vector of Date objects}

\item{time}{character time of day, defaults to '00:00:00'}

\item{tz}{character time zone, defaults to 'UTC'}

\item{format}{for character inputs, the format defaults to 'YYYY-mm-dd HH:MM:SS'}
}
\value{
POSIXct equivalent of input
}
\description{
Convert a Date, character, or POSIXlt to POSIXct
}
