% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.R
\name{assemble_gbm}
\alias{assemble_gbm}
\title{Given a list of one or more gbm.objects, retrieve a tibble that
summarizes the key diagnostic elements.}
\usage{
assemble_gbm(x)
}
\arguments{
\item{x}{a list of one or more gbm.objects}
}
\value{
a tibble of the summary values
}
\description{
Given a list of one or more gbm.objects, retrieve a tibble that
summarizes the key diagnostic elements.
}
