% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc_vector}
\alias{auc_vector}
\title{Compute AUC values ala presence-only data}
\usage{
auc_vector(f, v, thr = seq(from = 1, to = 0, by = -0.001), method = "fast")
}
\arguments{
\item{f}{a vector of forecasted ranked suitability values, NaNs and NAs will be removed}

\item{v}{a vector of forecasted values where there are presences, NaNs and NAs will be removed}

\item{thr}{a vector of threshold values}

\item{method}{character, if 'fast' (the default) use the fast implementation}
}
\value{
a list of
\itemize{
 \item{fpa, a vector of fpa (fractional predicted area)}
 \item{sensitive, a vector of sensitivity (1-omission rate)}
 \item{area, AUC}
 }
}
\description{
Compute AUC values ala presence-only data
}
