% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{compute_response}
\alias{compute_response}
\title{Compute maxent response}
\usage{
compute_response(x, var, at, data, cn, expand, fun, N)
}
\arguments{
\item{x}{model}

\item{var}{char, variable names}

\item{at}{see \code{\link[dismo]{response}}}

\item{data}{see \code{\link[dismo]{response}}}

\item{cn}{char, column names}

\item{expand}{see \code{\link[dismo]{response}}}

\item{fun}{see \code{\link[dismo]{response}}}

\item{N}{numeric, defines number of interpolates (rows)}
}
\description{
Compute maxent response
}
