% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{date_to_seasonname}
\alias{date_to_seasonname}
\title{Compute a season from POSIXct or Date values}
\usage{
date_to_seasonname(
  x,
  lut = c("DJF", "DJF", "MAM", "MAM", "MAM", "JJA", "JJA", "JJA", "SON", "SON", "SON",
    "DJF")
)
}
\arguments{
\item{x}{POSIXct or Date vector}

\item{lut}{charcater vector of length 12.  x is converted to month number
and is then used as a index into this look up table}
}
\value{
charcater vector of season names, same length as x
}
\description{
Compute a season from POSIXct or Date values
}
