% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{date_to_seasonprior}
\alias{date_to_seasonprior}
\alias{date_to_seasonprior.default}
\alias{date_to_seasonprior.POSIXct}
\alias{date_to_seasonprior.Date}
\title{Compute the prior season specified for each provided date.}
\usage{
date_to_seasonprior(x)

\method{date_to_seasonprior}{default}(x)

\method{date_to_seasonprior}{POSIXct}(x = Sys.time())

\method{date_to_seasonprior}{Date}(x = Sys.Date())
}
\arguments{
\item{x}{POSIXct or Date vector}
}
\value{
tibble with 5 columns of POSIXct vectors
 \itemize{
     \item{x original input dates}
     \item{DJF prior winter start dates}
     \item{MAM prior spring start dates}
     \item{JJA prior summer start dates}
     \item{SON prior autumn start dates}
}
}
\description{
Consider a spring (MAM) input date, the prior MAM will be 1 year before,
 while the prior DJF will be just three months before.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: for missing input

\item \code{POSIXct}: for POSIXct input

\item \code{Date}: for Date input
}}

