% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{doy_window}
\alias{doy_window}
\title{Compute the doy window given the day of year and window size}
\usage{
doy_window(x = 1, w = c(-5, 5), MAX = 366)
}
\arguments{
\item{x}{the day of year (should be 1-366 or 001-366)}

\item{w}{the window as 2 elements [days before, days after]}

\item{MAX}{the highest possible day number}
}
\value{
numeric doy vector
}
\description{
Given doy 100 with a window of c(-5,5) will yield -95,-96,...104,105
}
