% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.R
\name{extract_gbm}
\alias{extract_gbm}
\title{Extract a simple table of info from a single gbm.object}
\usage{
extract_gbm(x)
}
\arguments{
\item{x}{a gbm.object}
}
\value{
a tibble of summary results
}
\description{
Extract a simple table of info from a single gbm.object
}
