% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{images_to_gif}
\alias{images_to_gif}
\title{Create a gif animatation using ImageMagick's convert}
\usage{
images_to_gif(
  FF,
  output_filename = "animation.gif",
  convert_extra = "-loop 0 -delay 50"
)
}
\arguments{
\item{FF}{charcater vector of PNG filenames}

\item{output_filename}{the name of the output file, by default 'animation.gif'}

\item{convert_extra}{character of extra conver arguments.  Defaults to
"-loop 0 -delay = 50" for infinte looping and 0.5s per frame}
}
\value{
the numeric value returned by convert
}
\description{
Create a gif animatation using ImageMagick's convert
}
\seealso{
\url{http://www.imagemagick.org/Usage/anim_basics/}

\url{http://imagemagick.org/script/command-line-processing.php}
}
