% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{layers_extractPoints-deprecated}
\alias{layers_extractPoints-deprecated}
\alias{layers_extractPoints}
\title{Extract values from a multilayer Raster*}
\usage{
layers_extractPoints(R, pts)
}
\arguments{
\item{R}{a single or multilayer Raster* object}

\item{pts}{location info for points to be extracted. Must be a data frame
 or matrix with either set of the following columns. Note that layer (or z)
 may be either a layer index number or layer names.
\itemize{
     \item{'cell' and 'layer' (or 'z')}
     \item{'row', 'col' and 'layer' (or 'z')}
     \item{'lon', 'lat' and 'layer' (or 'z')}
     \item{'x', 'y' and 'layer' (or 'z')}
 }}
}
\value{
a vector of values
}
\description{
Extract values from a multilayer Raster*
}
