% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{layers_randomPoints}
\alias{layers_randomPoints}
\title{Select N non-NA random points from a mulitlayer Raster* object}
\usage{
layers_randomPoints(R, pts = NULL, N = 1000)
}
\arguments{
\item{R}{a multilayer Raster* object}

\item{pts}{xyz values for presence points [lon, lat, layer] - these 
 locations are avoided when sampling. Ignored if NULL.   This must have 
 either set of the following columns.  Note that layer (or z)
 may be either a layer index number or layer names.
\itemize{
     \item{'cell' and 'layer' (or 'z')} 
     \item{'row', 'col' and 'layer' (or 'z')}
     \item{'lon', 'lat' and 'layer' (or 'z')}
     \item{'x', 'y' and 'layer' (or 'z')}
 }}

\item{N}{the number of points to select}
}
\value{
data.frame of sampled points with the following
 \itemize{
     \item{lon}
     \item{lat}
     \item{col}
     \item{row}
     \item{cell}
     \item{layer}
     \item{value}
 }
}
\description{
Select N non-NA random points from a mulitlayer Raster* object
}

