% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxEntSummary.R
\name{maxent_assemble}
\alias{maxent_assemble}
\title{Read one or more maxent models stored in a path}
\usage{
maxent_assemble(version_path, drop_empty = TRUE)
}
\arguments{
\item{version_path}{path to directory containing one or more
maxent model directories}

\item{drop_empty}{logical if TRUE drop empty models}
}
\value{
a list of zero or more maxent models
}
\description{
Read one or more maxent models stored in a path
}

