% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{maxent_assemble_results}
\alias{maxent_assemble_results}
\title{Given a list of maxent models prepare a data frame of results}
\usage{
maxent_assemble_results(x, include = c("auc", "contribution"))
}
\arguments{
\item{x}{a list of one or more maxent models}

\item{include}{character vector fo one or more components to include in the result
- ignored for now}

\item{data.frame}{or NULL  Results for 'empty' models are excluded}
}
\description{
Given a list of maxent models prepare a data frame of results
}

