% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxEntSummary.R
\name{maxent_create_resultset}
\alias{maxent_create_resultset}
\title{Creates dataset of results of a set of maxEnt models. 
Dataset contains AUC and contributors of the set of maxEnt models
Drops models that have failed, then writes dataset to csv}
\usage{
maxent_create_resultset(models, filename)
}
\arguments{
\item{models}{one or more maxEnt models}

\item{filename}{absolute path to desired csv file}
}
\value{
dataframe
}
\description{
Creates dataset of results of a set of maxEnt models. 
Dataset contains AUC and contributors of the set of maxEnt models
Drops models that have failed, then writes dataset to csv
}

