% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{maxent_get_results}
\alias{maxent_get_results}
\title{Retrieve the results of a MaxEnt model by name}
\usage{
maxent_get_results(object, name)
}
\arguments{
\item{object}{a MaxEnt model}

\item{name}{the result name}
}
\value{
numeric, matrix or vector depending upon what is requested or NULL
}
\description{
Retrieve the results of a MaxEnt model by name
}
\examples{
\dontrun{
 auc <- maxent_get_results(object, 'AUC')
 contrib <- maxent_get_results(object, 'contribution')
}
}
