% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{maxent_get_varnames}
\alias{maxent_get_varnames}
\title{Retrieve the names of the input variables of a maxent model}
\usage{
maxent_get_varnames(object)
}
\arguments{
\item{object}{a MaxEnt model}
}
\value{
characater vector of variable names or NULL
}
\description{
Retrieve the names of the input variables of a maxent model
}

