% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{maxent_read_results}
\alias{maxent_read_results}
\title{Read MaxEnt results from a text file}
\usage{
maxent_read_results(filename)
}
\arguments{
\item{filename}{the fully qualified filename to read}
}
\value{
a matrix of results as per MaxEnt@results
}
\description{
Read MaxEnt results from a text file
}
