% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{mintersect}
\alias{mintersect}
\title{Compute the intersection of a list of vectors}
\usage{
mintersect(x)
}
\arguments{
\item{x}{a list of zero or more vectors}
}
\value{
a vector of the intersection or NULL is the input is empty
}
\description{
Compute the intersection of a list of vectors
}
