% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxEntSummary.R
\name{model_successful}
\alias{model_successful}
\title{Checks if a dismo model ran successfully}
\usage{
model_successful(model)
}
\arguments{
\item{model}{a model object such as MaxEnt, CliMap}
}
\value{
logical. If model ran successfully?
}
\description{
Checks if a dismo model ran successfully
}
