% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxEntSummary.R
\name{model_sucessful}
\alias{model_sucessful}
\title{checks if a dismo model ran sucessfully}
\usage{
model_sucessful(model)
}
\arguments{
\item{model}{a model object such as MaxEnt, CliMap}
}
\value{
logical. If model ran sucessfully?
}
\description{
checks if a dismo model ran sucessfully
}

