% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{munion}
\alias{munion}
\title{Compute the union of a list of vectors}
\usage{
munion(x)
}
\arguments{
\item{x}{a list of zero or more vectors}
}
\value{
a vector of the union or NULL is the input is empty
}
\description{
Compute the union of a list of vectors
}
