% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.R
\name{plot_gbm}
\alias{plot_gbm}
\title{Plot a gbm.object}
\usage{
plot_gbm(x, plot.folds = FALSE, name = NULL)
}
\arguments{
\item{x}{a gbm.object}

\item{plot.folds}{logical, if TRUE show the folds on the plot}

\item{name}{the name of the model to include in the plot, if NULL then
use x$gbm.call$gbm.y}
}
\value{
NULL invisibly
}
\description{
Plot a gbm.object
}
