% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{plot_model_summary}
\alias{plot_model_summary}
\title{Plot a three panel summary: contributions, AUC and counts}
\usage{
plot_model_summary(
  x = read_model_summary(),
  version = "v0.000",
  fauc = NULL,
  model_name = "model",
  auc_legend = "topright",
  count_legend = "topleft"
)
}
\arguments{
\item{x}{a tibble of summary data}

\item{version}{character, the model verison identifier}

\item{fauc}{tibble or NULL, if not NULL then
include the FAUC in the middle panel}

\item{model_name}{text, by default 'model' but can be anything you prefer}

\item{auc_legend}{text, location for legend in middle AUC/FAUC plot}

\item{count_legend}{text, location for legend in right hand counts plot}
}
\value{
nothing
}
\description{
Plot a three panel summary: contributions, AUC and counts
}
