% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{plot_response}
\alias{plot_response}
\title{Plot a response}
\usage{
plot_response(
  x,
  resp = maxent_response(plot = FALSE),
  data = NULL,
  rug = TRUE,
  ylim = c(0, 1),
  col = "#DF536B",
  lwd = 2,
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{maxent model}

\item{resp}{data frame of maxent response as per \code{\link{maxent_response}}}

\item{data}{data.frame or matrix of from the model or NULL}

\item{rug}{logical, if \code{TRUE} and \code{data} is present then add rug plot}

\item{ylim}{see \code{\link[dismo]{response}}}

\item{col}{see \code{\link[dismo]{response}}}

\item{lwd}{see \code{\link[dismo]{response}}}

\item{add}{see \code{\link[dismo]{response}}}

\item{...}{additional arguments for plotting}
}
\description{
Plot a response
}
