% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{read_maxent}
\alias{read_maxent}
\title{Read a dismo::MaxEnt model saved in a directory.}
\usage{
read_maxent(path)
}
\arguments{
\item{path}{fully qualified path to where the MexEnt model is saved}
}
\value{
a MaxEnt model object or NULL
}
\description{
Read a dismo::MaxEnt model saved in a directory.
}

