% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{read_model_summary}
\alias{read_model_summary}
\title{Read a summary file for a model family run - generally consists of a CSV file
with one row per model generated}
\usage{
read_model_summary(filename = "model_summary.csv")
}
\arguments{
\item{filename}{the filename with path}
}
\value{
either NULL or a tibble
}
\description{
Read a summary file for a model family run - generally consists of a CSV file
with one row per model generated
}
