% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{shannon_counts_index}
\alias{shannon_counts_index}
\title{Compute Shannon Index froma a table of counts}
\usage{
shannon_counts_index(x, base = exp(1))
}
\arguments{
\item{x}{table or matrix of observations.  If a matrix then, like a table,
columns are variables and rows are records.}

\item{base}{numeric, the base of the logarithm}
}
\value{
vector of Shannon index - one element for each record
}
\description{
Compute Shannon Index froma a table of counts
}
\seealso{
NIST \href{https://www.itl.nist.gov/div898/software/dataplot/refman2/auxillar/shannon.htm}{entry}
}
