% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{shannon_proportion_index}
\alias{shannon_proportion_index}
\title{Compute the Shannon index from a table of contributions}
\usage{
shannon_proportion_index(x, base = exp(1), offset = c(1e-05, NA)[1])
}
\arguments{
\item{x}{table or matrix of observations.  If a matrix then, like a table,
columns are variables and rows are records.}

\item{base}{numeric, the base of the logarithm}

\item{offset}{numeric, a small value to prevent taking the log of zero. If \code{offset} is
NA or less than or equal to zero then zero-valued elements are removed.
if \code{offset} greater than zero, then zero-valued elements in x are
replaced with this value.}
}
\value{
vector of Shannon index - one element for each record
}
\description{
Compute the Shannon index from a table of contributions
}
\seealso{
Wikipedia \href{https://en.wikipedia.org/wiki/Diversity_index#Shannon_index}{entry}
}
