% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{tomorrow}
\alias{tomorrow}
\title{Compute the next n day}
\usage{
tomorrow(x = today(), n = 1, ...)
}
\arguments{
\item{x}{POSIXct date as a starting date, generally \code{today()}}

\item{n}{the number of days ahead (or behind if negative)}

\item{...}{firther arguments for \code{today()} including form}
}
\value{
POSIXct
}
\description{
Compute the next n day
}
