% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{date_to_seasonprior}
\alias{date_to_seasonprior}
\title{Compute the prior season specified for each provided date.}
\usage{
date_to_seasonprior(x)
}
\arguments{
\item{x}{POSIXct or Date vector}
}
\value{
tibble with 5 columns of POSIXct vectors
 \itemize{
     \item{x original input dates}
     \item{DJF prior winter start dates}
     \item{MAM prior spring start dates}
     \item{JJA prior summer start dates}
     \item{SON prior autumn start dates}
}
}
\description{
Consider a spring (MAM) input date, the prior MAM will be 1 year before,
 while the prior DJF will be just three months before.
}
