% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{date_to_seasonstart}
\alias{date_to_seasonstart}
\title{Compute a season start from POSIXct or Date values}
\usage{
date_to_seasonstart(x)
}
\arguments{
\item{x}{POSIXct or Date vector}
}
\value{
POSIXct date that starts the season a date belongs to
}
\description{
Compute a season start from POSIXct or Date values
}

