% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{date_to_week}
\alias{date_to_week}
\title{Convert to as POSIXct to an 8D week number (1-46) or 7D week number (1-52)}
\usage{
date_to_week(x = today(), week_length = c(8, 7)[1])
}
\arguments{
\item{x}{POSIXct or Date vector}

\item{week_length}{numeric, the number of days per week}
}
\value{
numeric 1-46 or 1-52 week number
}
\description{
Convert to as POSIXct to an 8D week number (1-46) or 7D week number (1-52)
}
