% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismotools-package.R
\docType{package}
\name{dismotools}
\alias{dismotools}
\alias{dismotools-package}
\title{dismotools: Handy tools for the raster and dismo packages.}
\description{
Extra functionality for working with raster and dismo packages including model 
reading, model parameter extraction, sampling and extraction from multilayer 
raster objects.
}
