% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{maxent_response}
\alias{maxent_response}
\title{Compute the predictor response(s) for a maxent model.}
\usage{
maxent_response(
  x,
  var = NULL,
  at = median,
  range = "pa",
  expand = 10,
  data = NULL,
  fun = predict,
  N = 100
)
}
\arguments{
\item{x}{maxent model output}

\item{var}{character, one or more variables to compute response (defaults to all). See \code{\link[dismo]{response}}}

\item{at}{see \code{\link[dismo]{response}}}

\item{range}{see \code{\link[dismo]{response}}}

\item{expand}{see \code{\link[dismo]{response}}}

\item{data}{see \code{\link[dismo]{response}}}

\item{fun}{see \code{\link[dismo]{response}}}

\item{N}{numeric, the number of rows in the output.  Responses are interpolated at this number of intervals.}
}
\value{
tibble of sampled predictors and responses
}
\description{
Compute the predictor response(s) for a maxent model.
}
