% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{maxent_summary}
\alias{maxent_summary}
\title{Summarize a maxent model}
\usage{
maxent_summary(x, save_summary = FALSE, ...)
}
\arguments{
\item{x}{a MaxEnt model}

\item{save_summary}{logical, if TRUE then save to a text file as summary.txt}

\item{...}{further arguments for maxent_write_summary}
}
\value{
a list with summary information
\itemize{
 \item{path name of the model}
 \item{auc AUC value}
 \item{predictors vector of predictor names}
 \item{contrib data.frame of contributions by each predictor}
 \item{p_count number of presence points}
 \item{b_cound number of background points}
 }
}
\description{
Summarize a maxent model
}
