% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{maxent_write_summary}
\alias{maxent_write_summary}
\title{Write a maxent summary file}
\usage{
maxent_write_summary(x, filename = NULL)
}
\arguments{
\item{x}{list or data frame of maxent summary}

\item{filename}{optional filename to write to, by default file.path(x$path, "summary.txt")
or file.path(x$path, "summary.csv") if input is a data.frame}
}
\value{
the original list or data frame
}
\description{
Write a maxent summary file
}
