% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{read_maxent}
\alias{read_maxent}
\title{Read a dismo::MaxEnt model saved in a directory.}
\usage{
read_maxent(path)
}
\arguments{
\item{path}{fully qualified path to where the MexEnt model is saved
If multiple paths are provided then a list of models is returned.
You just have to keep track of this yourself so you aren't surprised.}
}
\value{
a MaxEnt model object, a list of Maxent models or NULL
}
\description{
Read a dismo::MaxEnt model saved in a directory.
}
