% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{seq_season}
\alias{seq_season}
\title{Generate a sequence of POSIXct or Date by season start dates}
\usage{
seq_season(from = as.Date("2003-01-01"), to = Sys.Date())
}
\arguments{
\item{from}{POSIXct or Date, the starting date}

\item{to}{POSIXct or Date, the ending date}
}
\value{
POSIXct or Date sequence
}
\description{
Generate a sequence of POSIXct or Date by season start dates
}
