% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{seq_season}
\alias{seq_season}
\title{Generate a sequence of POSIXct by season start dates}
\usage{
seq_season(from = as.POSIXct("2003-12-01 00:00:00", tz = "UTC"),
  to = today())
}
\arguments{
\item{from}{POSIXct, the starting date}

\item{to}{POSIXct, the ending date}
}
\value{
POSIXct sequence
}
\description{
Generate a sequence of POSIXct by season start dates
}

