% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{shannon_index}
\alias{shannon_index}
\title{Compute the Shannon index given a table of contributions}
\usage{
shannon_index(x, base = exp(1))
}
\arguments{
\item{x}{table or matrix of observations.  If a matrix then, like a table,
columns are variables and rows are records.}

\item{base}{numeric, the base of the logarithm}
}
\value{
vector of Shannon index - one element for each record
}
\description{
Compute the Shannon index given a table of contributions
}
\seealso{
Wikipedia \href{https://en.wikipedia.org/wiki/Diversity_index#Shannon_index}{entry}
}
