% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{today}
\alias{today}
\title{Compute today's date as POSIXct}
\usage{
today(x = as.POSIXct(paste(Sys.Date(), "00:00:00"), tz = tz),
  form = c("POSIXct", "\%Y-\%m-\%d", "week", "season")[1], tz = "UTC")
}
\arguments{
\item{x}{POSIXct or charcater date to be converted with as.POSIXct()}

\item{form}{the form of the date returned, possible a formatting pattern}
}
\value{
POSIXct
}
\description{
Compute today's date as POSIXct
}

