% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{today}
\alias{today}
\title{Compute today's date as POSIXct}
\usage{
today(x, form = c("POSIXct", "Date", "\%Y-\%m-\%d", "week", "season")[2])
}
\arguments{
\item{x}{Date or POSIXct or charcater date to be converted with as_POSIXct()}

\item{form}{the form of the date returned, possible a formatting pattern}
}
\value{
POSIXct or Date
}
\description{
Compute today's date as POSIXct
}
