% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{tomorrow}
\alias{tomorrow}
\title{Compute the next n day}
\usage{
tomorrow(x, n = 1)
}
\arguments{
\item{x}{Date or POSIXct, starting date, generally \code{today()}}

\item{n}{the number of days ahead (or behind if negative)}
}
\value{
POSIXct or Date
}
\description{
Compute the next n day
}
