% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{yesterday}
\alias{yesterday}
\title{Compute the prior n day.}
\usage{
yesterday(x, n = -1)
}
\arguments{
\item{x}{POSIXct date as a starting date, generally \code{today()}}

\item{n}{the number of days prior (or ahead if positive)}
}
\value{
POSIXct
}
\description{
A wrapper around \code{tomorrow()}
}
